requirements.txt
selenium
python-daemon
webdriver-manager